Added backend functionality for user handling and recipe adding, editing and deleting

+++ backend/controllers/recipes.js
@@ -1,7 +1,6 @@
 const recipesRouter = require('express').Router()
 const Recipe = require('../models/recipe')
 const User = require('../models/user')
-const crypto = require('crypto')
 
 recipesRouter.get('/', (req, res) => {
   Recipe.find({}).then(recipes => {
@@ -23,7 +22,7 @@ recipesRouter.get('/:id', async (req, res, next) => {
 })
 
 
-recipesRouter.post('/', async (req, res, next) => {
+recipesRouter.post('/:loginKEY', async (req, res, next) => {
   const body = req.body
 
   const recipe = new Recipe({
@@ -34,16 +33,16 @@ recipesRouter.post('/', async (req, res, next) => {
     steps: body.steps,
     servings: body.servings,
     timeToMake: body.timeToMake,
-    recipeID: crypto.randomBytes(20).toString('hex'),
     userID: body.userID,
   })
 
   try {
     const user = await User.findById(body.userID)
-    if (user) {
+    if (user && user.loginKEY === req.params.loginKEY) {
       const saved = await recipe.save()
       res.json(saved.toJSON())
     } else {
+      next(new Error('No user by that ID or not logged in'))
       res.status(404).end()
     }
 
@@ -52,35 +51,50 @@ recipesRouter.post('/', async (req, res, next) => {
   }
 })
 
-recipesRouter.delete('/:id', async (req, res, next) => {
+recipesRouter.delete('/:id&:userID&:loginKEY', async (req, res, next) => {
   try {
-    await Recipe.findByIdAndRemove(req.params.id)
-    res.status(204).end()
+    const user = await User.findById(req.params.userID)
+    if (user && user.loginKEY === req.params.loginKEY) {
+      await Recipe.findByIdAndRemove({_id: req.params.id, userID: user.id})
+      res.status(204).end()
+    } else {
+      next(new Error('No user by that ID or not logged in'))
+      res.status(404).end()
+    }
   } catch (exception) {
     next(exception)
   }
 })
 
-recipesRouter.put('/:id', (req, res, next) => {
+recipesRouter.put('/:id&:userID&:loginKEY', async (req, res, next) => {
   const body = req.body
+  try {
+    const user = await User.findById(req.params.userID)
+    if (user && user.loginKEY === req.params.loginKEY) {
+      try{
+        const recipe = {
+          category: body.category,
+          title: body.title,
+          description: body.description,
+          ingredients: body.ingredients,
+          steps: body.steps,
+          servings: body.servings,
+          timeToMake: body.timeToMake,
+        }
 
-  const recipe = {
-    category: body.category,
-    title: body.title,
-    description: body.description,
-    ingredients: body.ingredients,
-    steps: body.steps,
-    servings: body.servings,
-    timeToMake: body.timeToMake,
-    recipeID: crypto.randomBytes(20).toString('hex'),
+        const newRecipe = await Recipe.findByIdAndUpdate(req.params.id, recipe, { new: true})
+        res.json(newRecipe.toJSON())
+
+      } catch (exception) {
+        next(exception)
+      }
+    } else {
+      next(new Error('No user by that ID or not logged in'))
+      res.status(404).end()
+    }
+  } catch (exception) {
+    next(exception)
   }
-
-  Recipe.findByIdAndUpdate(req.params.id, recipe, { new: true})
-    .then(newRecipe => {
-      res.json(newRecipe.toJSON())
-    })
-    .catch(error => next(error))
 })
-
 
 module.exports = recipesRouter
